requirements.txt
imagehash
pytesseract
Pillow
numpy
